--- twilio_integration/fixtures/custom_field.json
@@ -1,108 +1,0 @@
-[
- {
-  "allow_in_quick_entry": 0,
-  "allow_on_submit": 0,
-  "bold": 0,
-  "collapsible": 0,
-  "collapsible_depends_on": null,
-  "columns": 0,
-  "default": null,
-  "depends_on": null,
-  "description": null,
-  "docstatus": 0,
-  "doctype": "Custom Field",
-  "dt": "Voice Call Settings",
-  "fetch_from": null,
-  "fetch_if_empty": 0,
-  "fieldname": "twilio_number",
-  "fieldtype": "Select",
-  "hidden": 0,
-  "hide_border": 0,
-  "hide_days": 0,
-  "hide_seconds": 0,
-  "ignore_user_permissions": 0,
-  "ignore_xss_filter": 0,
-  "in_global_search": 0,
-  "in_list_view": 0,
-  "in_preview": 0,
-  "in_standard_filter": 0,
-  "insert_after": "call_receiving_device",
-  "label": "Twilio Number",
-  "length": 0,
-  "mandatory_depends_on": null,
-  "modified": "2020-12-09 08:06:16.330327",
-  "name": "Voice Call Settings-twilio_number",
-  "no_copy": 0,
-  "non_negative": 0,
-  "options": null,
-  "parent": null,
-  "parentfield": null,
-  "parenttype": null,
-  "permlevel": 0,
-  "precision": "",
-  "print_hide": 0,
-  "print_hide_if_no_value": 0,
-  "print_width": null,
-  "read_only": 0,
-  "read_only_depends_on": null,
-  "report_hide": 0,
-  "reqd": 0,
-  "search_index": 0,
-  "translatable": 1,
-  "unique": 0,
-  "width": null
- },
- {
-  "allow_in_quick_entry": 0,
-  "allow_on_submit": 0,
-  "bold": 0,
-  "collapsible": 0,
-  "collapsible_depends_on": null,
-  "columns": 0,
-  "default": null,
-  "depends_on": "eval: doc.channel==='WhatsApp'",
-  "description": "To use WhatsApp for Business, initialize <a href=\"#Form/Twilio Settings\">Twilio Settings</a>.",
-  "docstatus": 0,
-  "doctype": "Custom Field",
-  "dt": "Notification",
-  "fetch_from": null,
-  "fetch_if_empty": 0,
-  "fieldname": "twilio_number",
-  "fieldtype": "Link",
-  "hidden": 0,
-  "hide_border": 0,
-  "hide_days": 0,
-  "hide_seconds": 0,
-  "ignore_user_permissions": 0,
-  "ignore_xss_filter": 0,
-  "in_global_search": 0,
-  "in_list_view": 0,
-  "in_preview": 0,
-  "in_standard_filter": 0,
-  "insert_after": "slack_webhook_url",
-  "label": "Twilio Number",
-  "length": 0,
-  "mandatory_depends_on": "eval: doc.channel==='WhatsApp'",
-  "modified": "2020-10-26 11:05:12.231781",
-  "name": "Notification-twilio_number",
-  "no_copy": 0,
-  "non_negative": 0,
-  "options": "Communication Medium",
-  "parent": null,
-  "parentfield": null,
-  "parenttype": null,
-  "permlevel": 0,
-  "precision": "",
-  "print_hide": 0,
-  "print_hide_if_no_value": 0,
-  "print_width": null,
-  "read_only": 0,
-  "read_only_depends_on": null,
-  "report_hide": 0,
-  "reqd": 0,
-  "search_index": 0,
-  "translatable": 0,
-  "unique": 0,
-  "width": null
- }
-]
