+++ twilio_integration/fixtures/property_setter.json
@@ -1,67 +1,33 @@
 [
  {
-  "default_value": null,
-  "doc_type": "Communication Medium",
-  "docstatus": 0,
-  "doctype": "Property Setter",
-  "doctype_or_field": "DocField",
-  "field_name": "communication_channel",
-  "modified": "2020-10-26 10:39:10.230657",
-  "name": "Communication Medium-communication_channel-options",
-  "parent": null,
-  "parentfield": null,
-  "parenttype": null,
-  "property": "options",
-  "property_type": "Text",
-  "row_name": null,
-  "value": "\nExotel\nTwilio"
- },
- {
+  "creation": "2025-04-30 21:27:29.017511",
   "default_value": null,
   "doc_type": "Notification",
   "docstatus": 0,
   "doctype": "Property Setter",
   "doctype_or_field": "DocField",
   "field_name": "channel",
+  "is_system_generated": 1,
   "modified": "2020-11-01 11:05:40.052493",
+  "module": null,
   "name": "Notification-channel-options",
-  "parent": null,
-  "parentfield": null,
-  "parenttype": null,
   "property": "options",
   "property_type": "Text",
   "row_name": null,
   "value": "Email\nSlack\nSystem Notification\nSMS\nWhatsApp"
  },
  {
-  "default_value": null,
-  "doc_type": "Notification",
-  "docstatus": 0,
-  "doctype": "Property Setter",
-  "doctype_or_field": "DocField",
-  "field_name": "column_break_5",
-  "modified": "2020-10-25 11:06:45.707981",
-  "name": "Notification-column_break_5-depends_on",
-  "parent": null,
-  "parentfield": null,
-  "parenttype": null,
-  "property": "depends_on",
-  "property_type": "Data",
-  "row_name": null,
-  "value": "eval:in_list(['Email', 'SMS', 'WhatsApp'], doc.channel)"
- },
- {
+  "creation": "2025-04-30 21:27:29.043581",
   "default_value": null,
   "doc_type": "Communication Medium",
   "docstatus": 0,
   "doctype": "Property Setter",
   "doctype_or_field": "DocField",
   "field_name": "communication_medium_type",
+  "is_system_generated": 1,
   "modified": "2020-10-24 15:04:25.859806",
+  "module": null,
   "name": "Communication Medium-communication_medium_type-options",
-  "parent": null,
-  "parentfield": null,
-  "parenttype": null,
   "property": "options",
   "property_type": "Text",
   "row_name": null,
