+++ twilio_integration/twilio_integration/doctype/whatsapp_message/whatsapp_message.py
@@ -2,12 +2,13 @@
 # For license information, please see license.txt
 
 import frappe
+from frappe import _
 from frappe.model.document import Document
-from six import string_types
 from frappe.utils.password import get_decrypted_password
 from frappe.utils import get_site_url
-from frappe import _
 from ...twilio_handler import Twilio
+import json
+
 
 class WhatsAppMessage(Document):
 	def send(self):
@@ -25,7 +26,7 @@ class WhatsAppMessage(Document):
 
 		except Exception as e:
 			self.db_set('status', "Error")
-			frappe.log_error(title = _('Twilio WhatsApp Message Error'),message=e)
+			self.log_error(title=_('Twilio WhatsApp Message Error'), message=e)
 
 	def get_message_dict(self):
 		args = {
@@ -40,53 +41,67 @@ class WhatsAppMessage(Document):
 		return args
 
 	@classmethod
-	def send_whatsapp_message(self, receiver_list, message, doctype, docname, notification_type=None, media=None):
+	def send_whatsapp_message(
+		cls,
+		receiver_list,
+		message,
+		doctype,
+		docname,
+		notification_type=None,
+		media=None,
+		communication=None
+	):
 		from frappe.email.doctype.notification.notification import get_doc_for_notification_triggers
 
-		if isinstance(receiver_list, string_types):
-			receiver_list = loads(receiver_list)
+		if are_whatsapp_messages_muted():
+			frappe.msgprint(_("WhatsApp is muted"))
+			return
+
+		if isinstance(receiver_list, str):
+			receiver_list = json.loads(receiver_list)
 			if not isinstance(receiver_list, list):
 				receiver_list = [receiver_list]
 		doc = get_doc_for_notification_triggers(doctype, docname)
 
-		run_before_send_method(doc=doc,notification_type=notification_type)
-
-		if are_whatsapp_messages_muted():
-			frappe.msgprint(_("Whatsapp is muted"))
-			return
+		run_before_send_method(doc=doc, notification_type=notification_type)
 
 		for rec in receiver_list:
-			message = self.store_whatsapp_message(rec, message, doctype, docname)
-			message.send()
+			wa_msg = cls.store_whatsapp_message(rec, message,
+				doctype=doctype, docname=docname, media=media, communication=communication)
+			wa_msg.send()
 
 		run_after_send_method(doctype=doctype, docname=docname, notification_type=notification_type)
 
-	def store_whatsapp_message(to, message, doctype=None, docname=None, media=None):
+	@classmethod
+	def store_whatsapp_message(cls, to, message, doctype=None, docname=None, media=None, communication=None):
 		sender = frappe.db.get_single_value('Twilio Settings', 'whatsapp_no')
 		wa_msg = frappe.get_doc({
-				'doctype': 'WhatsApp Message',
-				'from_': 'whatsapp:{}'.format(sender),
-				'to': 'whatsapp:{}'.format(to),
-				'message': message,
-				'reference_doctype': doctype,
-				'reference_document_name': docname,
-				'media_link': media
-			}).insert(ignore_permissions=True)
+			'doctype': 'WhatsApp Message',
+			'from_': 'whatsapp:{}'.format(sender),
+			'to': 'whatsapp:{}'.format(to),
+			'message': message,
+			'reference_doctype': doctype,
+			'reference_document_name': docname,
+			'media_link': media,
+			'communication': communication,
+		}).insert(ignore_permissions=True)
 
 		return wa_msg
 
+
 def incoming_message_callback(args):
 	wa_msg = frappe.get_doc({
-			'doctype': 'WhatsApp Message',
-			'from_': args.From,
-			'to': args.To,
-			'message': args.Body,
-			'profile_name': args.ProfileName,
-			'sent_received': args.SmsStatus.title(),
-			'id': args.MessageSid,
-			'send_on': frappe.utils.now(),
-			'status': 'Received'
-		}).insert(ignore_permissions=True)
+		'doctype': 'WhatsApp Message',
+		'from_': args.From,
+		'to': args.To,
+		'message': args.Body,
+		'profile_name': args.ProfileName,
+		'sent_received': args.SmsStatus.title(),
+		'id': args.MessageSid,
+		'send_on': frappe.utils.now(),
+		'status': 'Received'
+	}).insert(ignore_permissions=True)
+
 
 def run_before_send_method(doc=None, notification_type=None):
 	from frappe.email.doctype.notification.notification import run_validate_notification
@@ -98,11 +113,13 @@ def run_before_send_method(doc=None, notification_type=None):
 		if not validation:
 			frappe.throw(_("{0} Notification Validation Failed").format(notification_type))
 
+
 def run_after_send_method(doctype=None, docname=None, notification_type=None):
 	from frappe.core.doctype.notification_count.notification_count import add_notification_count
 
 	if doctype and docname and notification_type:
 		add_notification_count(doctype, docname, notification_type, 'WhatsApp')
+
 
 def are_whatsapp_messages_muted():
 	from frappe.utils import cint
@@ -114,4 +131,3 @@ def are_whatsapp_messages_muted():
 		pass
 
 	return cint(frappe.conf.get("mute_whatsapp") or 0) or False
-
